+++ include/fluxgl/audio/source.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <miniaudio.h>
+
+namespace fluxgl {
+    struct Source {
+        ma_sound sound;
+        float volume = 1.0f;
+        float loop = false;
+    };
+}