+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2026-02-25
+
+### Added
+- Added a basic audioEngine (load / play / start) based on miniaudio lib 
+- Implemented the ECS Audio_System with AudioSource and AudioListener components for audio capabilities
+
+### Fixed
+- Fixed a rotation compute error that made the directional lights not work properly
+
 ## [0.3.2] - 2026-02-24
 
 ### Fixed